NoticiaScraper/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "lxml>=5.4.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "requests>=2.32.4",
    "schedule>=1.2.2",
    "selenium>=4.34.2",
    "selenium-stealth>=1.0.6",
    "trafilatura>=2.0.0",
    "webdriver-manager>=4.0.2",
]
